+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## CHANGES OR IMPROVEMENTS
 
-Renamed the `abc` function argument `ninit` to `n_init`.
+* Renamed the `abc` function argument `ninit` to `n_init`.
+
+* Renamed the `abc` function argument `npart` to `n_particles`.
 
 # SimInf 9.8.1 (2024-06-21)